server/requirements.txt
channels
channels_redis
django
pycryptodome
